+++ backend/movies/views/__init__.py
@@ -1,0 +1,17 @@
+from .actor_search_view import ActorSearchAPIView
+from .crew_search_view import CrewSearchAPIView
+from .movie_credits_view import MovieCreditsAPIView
+from .movie_details_view import MovieMetadataAPIView
+from .movie_search_view import MovieSearchAPIView
+from .person_details_view import PersonMetadataAPIView
+from .popular_movies_view import PopularMoviesAPIView
+
+__all__ = [
+    'ActorSearchAPIView',
+    'CrewSearchAPIView',
+    'MovieCreditsAPIView',
+    'MovieMetadataAPIView',
+    'MovieSearchAPIView',
+    'PersonMetadataAPIView',
+    'PopularMoviesAPIView',
+]
